fix search bug
add condition in paggination.js for search
